fix: break hw↔simulation import cycle by moving NamedObject to waveflow.named
SimObj imported NamedObject from waveflow.hw.named; importing any waveflow.hw.*
leaf eagerly runs hw/__init__.py, which pulls the interface layer back down onto
the half-initialized simobj module — a layering inversion that made a cold
`from waveflow.simulation.simobj import SimObj` raise a circular ImportError.

NamedObject is a pure-stdlib leaf, so move it to the neutral top-level
waveflow/named.py (the package __init__ is empty, re-triggering nothing) and
update the 4 import sites. Both import directions now resolve cold. Drop the
`import waveflow.hw` workaround from the SimObj-toy doc snippet, which now
imports cleanly.
